Update TF Datasets template to remove jQuery2.
We have a request to remove uses of jQuery2. Vanilla JS is mature enough that this can be re-written succintly without an external dependency, so that's what I've done.

All of the TF Docs usage of jQ2 stems from this template, so once the docs are regenerated, TF should be in good standing.

PiperOrigin-RevId: 426029849